requirements.txt
iterative-stratification==0.1.6
numpy==1.17.3
Pillow==6.2.0
torch==1.3.0
torchvision==0.4.1
pandas==0.25.1
imgaug==0.3.0
scikit-learn==0.19.0
opencv-python==4.1.1.26
efficientnet-pytorch==0.5.0
tqdm==4.40.2
gdown==3.13.0
streamlit==0.83.0